// client/src/models/generated/DealsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Deal } from "./Deal";

export interface DealsResponse { chain_id: string, zip_code: string, valid_from: string | null, valid_to: string | null, deals: Array<Deal>, cached: boolean, }